Gems update #1
Gems are now in the game! The more you carry, the heavier you are. Left-Click (or Q) to throw them and Right-Click (or E) to pick them up.
What doesn't work: indicator showing number of gems doesn't work, I'll fix that, and maybe other stuff too I'll continue working on it